Reduce the input/analysis queue file open polling attempts
An attempt should only be made to open the next queue file when
the current file is greater than or equal to the output_size.

Warning: increasing the output_size cfg option without clearing
the output queues can cause the readers to stall waiting for an
already finalized file to grow to the new size.